ドメイン層にUserモデルを追加しました。またファイル名がuser_name.rsからuser.rsに変更したことに伴いこのファイルを参照している部分のパスを変更しました。
