+++ Cargo.toml
@@ -17,3 +17,4 @@ tower-http = { version = "0.5.0", features = ["add-extension", "trace"] }
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 dotenvy = "0.15.7"
+axum-login = "0.17.0"
