+++ src/domain/models/article.rs
@@ -2,7 +2,7 @@ use bson::oid::ObjectId;
 use serde::{Deserialize, Serialize};
 use chrono::{DateTime, Utc};
 
-use crate::domain::models::user_name::UserName;
+use crate::domain::models::user::UserName;
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct Article {
     #[serde(rename = "_id")]
