+++ src/domain/models/mod.rs
@@ -1,2 +1,2 @@
 pub mod article;
-pub mod user_name;
+pub mod user;
