+++ src/usecase/article_usecase.rs
@@ -1,7 +1,7 @@
 use async_trait::async_trait;
 use mongodb::bson::oid::ObjectId;
 
-use crate::domain::{models::{article::Article, user_name::UserName}, repositorys::article_repository::ArticleRepository};
+use crate::domain::{models::{article::Article, user::UserName}, repositorys::article_repository::ArticleRepository};
 
 #[derive(Clone)]
 pub struct ArticleUsecase<T: ArticleRepository + Clone> {
